Revised tracing code (wwxwrxtest)

+++ wwxwrxtest/WebRole1/Default.aspx.cs
@@ -15,19 +15,23 @@ namespace WebRole1
     public partial class Default : System.Web.UI.Page
     {
         string winwrapTraceFileName = AppDomain.CurrentDomain.GetData("DataDirectory") + @"\winwrap.txt";
+        static object lock_ = new object();
 
         [System.Diagnostics.ConditionalAttribute("DEBUG")]
         void AppendToLog(string text)
         {
-            using (System.IO.StreamWriter sw = System.IO.File.AppendText(winwrapTraceFileName))
+            lock (lock_)
             {
-                sw.WriteLine(DateTime.UtcNow.ToString() + " [" + System.Threading.Thread.CurrentThread.ManagedThreadId + "] " + text);
+                using (System.IO.StreamWriter sw = System.IO.File.AppendText(winwrapTraceFileName))
+                {
+                    sw.WriteLine(DateTime.UtcNow.ToString() + " [" + System.Threading.Thread.CurrentThread.ManagedThreadId + "] " + text);
+                }
             }
         }
 
         protected void Page_Load(object sender, EventArgs e)
         {
-            if (System.IO.File.Exists(winwrapTraceFileName))
+            if (System.IO.File.Exists(winwrapTraceFileName) && Request["ResetTrace"] == "yes")
                 System.IO.File.Delete(winwrapTraceFileName);
 
             AppendToLog("Creating BasicNoUIObj");
@@ -38,19 +42,19 @@ namespace WebRole1
 #endif
                 basicNoUIObj.Secret = new Guid(Utils.GetPatternString("wwxwrxtest", "Guid[(]\"(.*)\"[)]"));
                 //basicNoUIObj.Secret = Guid.NewGuid();
-                basicNoUIObj.Initialize();
-#if DEBUG
                 try
                 {
+                    basicNoUIObj.Initialize();
                     AppendToLog("Turn on BasicNoUIObj tracing");
+#if DEBUG
                     basicNoUIObj.Trace(WinWrap.Basic.TraceConstants.All);
+#endif
                 }
                 catch (Exception ex)
                 {
                     AppendToLog("Exception: " + ex.Message + "\r\n" + ex.StackTrace.ToString());
                     throw ex;
                 }
-#endif
 
                 Button1.Text = basicNoUIObj.Evaluate("4+3");
             }
